% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph.R
\name{add_exit}
\alias{add_exit}
\title{Add an exit vertice to an existing network}
\usage{
add_exit(df_edges, name_to = "exit")
}
\arguments{
\item{df_edges}{data.frame of existing network with columns `from`, `to`, `weight`, `level`, `set`.}

\item{name_to}{string, name of the exit/sink vertice. Default is "exit".}
}
\value{
data.frame of existing network with columns `from`, `to`, `weight`, `level`, `set`.
}
\description{
The function adds exit vertice to an existing network with specified edges.
}
\details{
## Weights:
Weights are set to the edges in a direction from the left to the right from the side of the newly added vertices.
}
\examples{

network <- create_source(set = c(2,2),
                         weight = c(1)) \%>\%
 add_exit(name_to="LAST VERTICE")
plot_network(network)

network2 <- create_source(c(2,2),
                          c(1,2)) \%>\%
 add_fully_connected(2,
                     1) \%>\%
 add_exit()
plot_network(network2)

}
