% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_helpers.R
\name{add_level_text}
\alias{add_level_text}
\title{Add text to df_text at level}
\usage{
add_level_text(input, df_new_text, x_coord = 0, snap = NULL)
}
\arguments{
\item{input}{a `CVCS_simulation` object get from a \code{\link{simulate_flow}} or directly df_text with columns `snap`, `level`, `x`, `y`, `text`.}

\item{df_new_text}{data.frame, with columns `level` and `text`, `level` should reference to levels and `y` coordinate is computed based on this.}

\item{x_coord}{numeric, this is the `x` coordinate.}

\item{snap}{numeric vector, this variable can be used to put the text only in certain timesteps (snaps).}
}
\value{
an updated `CVCS_simulation` object in case of `CVCS_simulation` object input and updated data.frame in case of data.frame.
}
\description{
Add text to the text object for simulation at each level.
}
\examples{

df_network <- create_source(set = c(1,1),
                             weight = 3) \%>\%
 add_fully_connected(weight = c(1,1,2,2),
                     set = c(10,1,1,1)) \%>\%
 add_level_set(weight = list(c(2,1),
                             4,
                             6),
               previous_set = list(c(1,2),
                                   3,
                                   c(3,4)),
               set = c(2, 3, 1)) \%>\%
 add_exit()

res <- simulate_flow(df_network)

net_levels <- get_network_levels(df_network)
net_text <- 1:length(net_levels)

new_text <- data.frame(level = net_levels,
                       text = net_text)

res <- add_level_text(res,
                      new_text,
                      x_coord = 6.8)

animate_simulation(df_network,
                   res[[1]],
                   res[[2]],
                   snap = 0)

}
