% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph.R
\name{create_source}
\alias{create_source}
\title{Create edges from source}
\usage{
create_source(set, weight, name_from = "source", name_to = "1_")
}
\arguments{
\item{set}{integer vector, each number represents a number of vertices in a set}

\item{weight}{integer vector, \itemize{
\item{}{if single number is provided the same weight is set for all the edges,}
\item{}{if the number of elements in a vector equals to the `length(set)` then the weight is the same for each set,}
\item{}{if the number of elements equals to the `sum(set)` then each edge has unique weight.}}}

\item{name_from}{string, name of the source vertice. Default is "source".}

\item{name_to}{string, prefix for the names of vertices on the next level. Each vertice will have number appended to it for the uniqueness of names. Default is "1_".}
}
\value{
data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`.
}
\description{
This function creates data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`
starting from the "source" vertice. It is an initial function to create a network.
}
\details{
## Weights:
Weights are set to the edges in a direction from the left to the right.
}
\examples{

network <- create_source(set = c(2,1),
                         weight = c(5,2))
plot_network(network)

network2 <- create_source(set = c(2,1),
                          weight = 1)
plot_network(network2)

network3 <- create_source(set = c(2,1),
                          weight = c(1,5,10),
                          name_from = "entrance",
                          name_to = "registration")
plot_network(network3)

}
