% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cvlm}
\alias{cvlm}
\title{Check linear models with cross validation}
\usage{
cvlm(y, x, subsize, iter = 100)
}
\arguments{
\item{y}{dependent variable.}

\item{x}{design matrix (should include constant if it is needed).}

\item{subsize}{sample size of subsampling.}

\item{iter}{number of iterations.}
}
\value{
The function returns a list with following components.
\item{rmse}{root mean squares of forecast errors for all iterations.}
\item{mae}{mean absolute forecast errors for all iterations.}
}
\description{
The function checks linear models with cross-validation (out-of-sample prediction).
}
\references{
Tsay, R. and Chen, R. (2019). Nonlinear Time Series Analysis. Wiley, New Jersey.
}
