% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{F.test}
\alias{F.test}
\title{F Test for Nonlinearity}
\usage{
F.test(x, order, thres = 0)
}
\arguments{
\item{x}{time series.}

\item{order}{AR order.}

\item{thres}{threshold value.}
}
\value{
The function outputs the test statistic and its p-value, and return a list with components:
\item{test.stat}{test statistic.}
\item{p.value}{p-value.}
\item{order}{AR order.}
}
\description{
Compute the F-test statistic for nonlinearity
}
\examples{
y=rnorm(100)
F.test(y,2,0)
}
