% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flow.R
\name{get_level_throughput}
\alias{get_level_throughput}
\title{Calculate the total flow through the individual levels of the network.}
\usage{
get_level_throughput(network, constrain = TRUE)
}
\arguments{
\item{network}{data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`}

\item{constrain}{bool deciding on constraining the maximum possible flow in a level to the maximum flow in the previous level.}
}
\value{
levels_throughtput data.frame containing the maximum flow in network levels with the columns `level_id`, `throughtput`
}
\description{
This function calculates the total flow through the individual
network levels in the defined network.
}
\examples{
network <- create_source(set = c(3,2,5), weight = 2.5) \%>\%
add_level_set(weight = c(2,4,1), previous_set = c(3,2,1),set = c(2,3, 1)) \%>\%
  add_level_set(weight = c(2,2,2),previous_set = c(1,2,3),set = c(1,1,1)) \%>\%
  add_fully_connected(weight = 5,set = 4) \%>\%
  add_exit()

level_throughput <- get_level_throughput(network)

}
