% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_helpers.R
\name{get_network_levels_coord}
\alias{get_network_levels_coord}
\title{Get levels coord}
\usage{
get_network_levels_coord(df_network, df_vertices = NULL)
}
\arguments{
\item{df_network}{data.frame with network defined by edges. It should contain columns `from`, `to`, `weight`, `level`.}

\item{df_vertices}{data.frame, with description of vertices get by the with columns `vertice`, `level`, `x`, `y`. It is an optional parameter since usually this is computed from `df_network`.}
}
\value{
vector with levels of the network
}
\description{
Get levels coord from network edge definition.
}
\examples{
df_network <- create_source(set = c(1,1),
                             3) \%>\%
 add_exit()

res <- simulate_flow(df_network)

net_levels <- get_network_levels_coord(df_network)

}
