% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flow.R
\name{get_node_load}
\alias{get_node_load}
\title{Calculate the load of nodes in the network}
\usage{
get_node_load(network, constrain = TRUE)
}
\arguments{
\item{network}{data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`.}

\item{constrain}{bool deciding on constraining the maximum possible flow in a level to the maximum flow in the previous nodes.}
}
\value{
nodes data.frame with the definition of nodes by the columns `id`, `input_people`, `output_people`, `node_load`
}
\description{
This function calculates the maximum possible input and output
from the nodes and its load in this flow using the edges in the network.
}
\examples{
network <- create_source(set = c(3,2,5), weight = 2.5) \%>\%
add_level_set(weight = c(2,4,1), previous_set = c(3,2,1),set = c(2,3, 1)) \%>\%
  add_level_set(weight = c(2,2,2),previous_set = c(1,2,3),set = c(1,1,1)) \%>\%
  add_fully_connected(weight = 5,set = 4) \%>\%
  add_exit()

node_load <- get_node_load(network)

}
