% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hfDummy}
\alias{hfDummy}
\title{Create Dummy Variables for High-Frequency Intraday Seasonality}
\usage{
hfDummy(int = 1, Fopen = 10, Tend = 10, days = 1, pooled = 1,
  skipmin = 0)
}
\arguments{
\item{int}{length of time interval in minutes.}

\item{Fopen}{number of dummies/intervals from the market open.}

\item{Tend}{number of dummies/intervals to the market close.}

\item{days}{number of trading days in the data.}

\item{pooled}{a logical value indicating whether the data are pooled.}

\item{skipmin}{the number of minites omitted from the opening.}
}
\description{
Create dummy variables for high-frequency intraday seasonality.
}
\examples{
x=hfDummy(5,Fopen=4,Tend=4,days=2,skipmin=15)
}
