% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{make_event}
\alias{make_event}
\title{Make/add event}
\usage{
make_event(
  event_list = list(),
  value,
  column,
  snap,
  table = "vertice",
  vertices = NA,
  state = NA,
  vertex_color = NA,
  from = NA,
  to = NA
)
}
\arguments{
\item{event_list}{list with event description created previously by this function, default is empty list and it does not need to be filled when creating first event.}

\item{value}{value which shall be replaced.}

\item{column}{which column of a table should be affected.}

\item{snap}{numeric vector, this  (snaps).}

\item{table}{string, table name which should be affected. It should be 'vertice' for changing vertice description and 'edge' for changing the network architecture.}

\item{vertices}{when table is 'vertice' a list of vertices which shall be affected.}

\item{state}{when table is 'vertice' this variable affect its state column. This is a convenient variable to facilitate changing several columns with one command.}

\item{vertex_color}{when table is 'vertice' this variable affect its color column. This is a convenient variable to facilitate changing several columns with one command.}

\item{from}{string, when table is 'edge' a `from` is used to subset the edges by the from column. If not used the subsetting is done solely on the to variable.}

\item{to}{string, when table is 'edge' a `to` is used to subset the edges by the to column. If not used the subsetting is done solely on the from variable.}
}
\value{
list with events definition.
}
\description{
Make or add event
}
\examples{


df_edges <- create_source(set = c(1,2,1),
                            weight = c(5,2,10)) \%>\%
 add_level_set(weight = c(2,1,1),
               previous_set = list(1,2,3),
               set = c(2,2,1)) \%>\%
 add_level_set(weight = c(4,2),
               previous_set = list(c(1,2), 3),
               set = c(2,2)) \%>\%
 add_fully_connected(set = c(1,1,1),
                     weight = 2) \%>\%
 add_level_set(weight = c(4,2,1),
               previous_set = list(1,2,3),
               set = c(2,1,1)) \%>\%
 add_exit()

l_events <- make_event(value = 300000,
                      table = "vertice",
                       column = "remaining_time",
                      vertices = c("3_3", "3_4"),
                      vertex_color = 2,
                      state = "off",
                      snap = 20) \%>\%
 make_event(value = 1000,
            table = "edge",
            column = "weight",
            # from = "2_2",
            to = "3_1",
            snap = 10) \%>\%
 make_event(value = 1,
            column = "remaining_time",
            vertices = c("3_3", "3_4"),
            vertex_color = 1,
            state = "on",
            snap = 40) \%>\%
 make_event(value = 1,
            column = "remaining_time",
            vertices = c("3_1"),
            vertex_color = 1,
            state = "on",
            snap = 25) \%>\%
 make_event(value = 1,
            table = "edge",
            column = "weight",
            # from = "2_2",
            to = "3_1",
            snap = 20)

res <- simulate_flow(df_edges = df_edges,
                     l_events = l_events,
                     timestep = 1,
                     input_people = 40)

}
