% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_set_list}
\alias{make_set_list}
\title{Expand set list}
\usage{
make_set_list(set, last_level_sets_lengths = NULL)
}
\arguments{
\item{set}{numeric vector with set definition.}

\item{last_level_sets_lengths}{numeric vector with the number of vertices of previous sets.}
}
\description{
Expand set list
}
