% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_weight_list}
\alias{make_weight_list}
\title{Expand weights}
\usage{
make_weight_list(set, weight, last_level_sets_lengths = NULL)
}
\arguments{
\item{set}{numeric vector with set definition.}

\item{weight}{numeric vector with a weight definition.}

\item{last_level_sets_lengths}{numeric vector with the number of vertices of previous sets.}
}
\description{
This function prepares
}
