% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{MSM.fit}
\alias{MSM.fit}
\title{Fitting Univariate Autoregressive Markov Switching Models}
\usage{
MSM.fit(y, p, nregime = 2, include.mean = T, sw = NULL)
}
\arguments{
\item{y}{a time series.}

\item{p}{AR order.}

\item{nregime}{the number of regimes.}

\item{include.mean}{a logical value for including constant terms.}

\item{sw}{logical values for whether coefficients are switching. The length of \code{sw} has to be equal to the number of coefficients in the model plus include.mean.}
}
\value{
\code{MSM.fit} returns an object of class code{MSM.lm} or \code{MSM.glm}, depending on the input model.
}
\description{
Fit autoregressive Markov switching models to a univariate time series using the package MSwM.
}
