% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{operation_over_set}
\alias{operation_over_set}
\title{Do operation over level of input data.frame}
\usage{
operation_over_set(current_set, df_edges, max_level, fnc, ...)
}
\arguments{
\item{current_set}{numeric, set id which should be worked on}

\item{df_edges}{data.frame, edge description of the network}

\item{max_level}{numeric, level which should be subsetted}

\item{fnc}{function, this is a function which should be made over the subset.}

\item{...}{additional arguments for the `fnc`}
}
\description{
This function subset the vertices from the max level of the network and then do a operation on them defined by `fnc`.
}
