% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{p_cfar_part}
\alias{p_cfar_part}
\title{Partial Curve Prediction of CFAR Processes}
\usage{
p_cfar_part(model, f, new.obs)
}
\arguments{
\item{model}{CFAR model.}

\item{f}{the functional time series data.}

\item{new.obs}{the given first \code{n} observations.}
}
\value{
The function returns a prediction of the CFAR process.
}
\description{
Partial prediction for CFAR processes. t curves are given and we want to predit the curve at time t+1, but we know the first n observations in the curve, to predict the n+1 observation.
}
\examples{
phi_func= function(x)
{
 	return(dnorm(x,mean=0,sd=0.1))
}
y=g_cfar1(100,5,phi_func)
f_grid=y$cfar
index=seq(1,1001,by=10)
f=f_grid[,index]
est=est_cfar(f,1)
pred=p_cfar_part(est,f,0)
}
\references{
Liu, X., Xiao, H., and Chen, R. (2016) Convolutional autoregressive models for functional time series. \emph{Journal of Econometrics}, 194, 263-282.
}
