% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Tsay}
\alias{Tsay}
\title{Tsay Test for Nonlinearity}
\usage{
Tsay(y, p = 1)
}
\arguments{
\item{y}{time series.}

\item{p}{AR order.}
}
\value{
The function outputs the F statistic, p value, and the degrees of freedom. The null hypothsis is there is no nonlinearity.
}
\description{
Perform Tsay (1986) nonlinearity test.
}
\examples{
y=MSM.sim(100,c(1,1),0.7,-0.5,c(0.5,0.6),c(1,1),c(0,0),500)
Tsay(y$series,1)
}
\references{
Tsay, R. (1986) Nonlinearity tests for time series. \emph{Biometrika} \strong{73}(2), 461-466.
}
