% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.r
\name{wrap.SMC}
\alias{wrap.SMC}
\title{Sequential Monte Carlo Using Sequential Importance Sampling for Stochastic Volatility Models}
\usage{
wrap.SMC(par.natural, yy, mm, setseed = T, resample = T)
}
\arguments{
\item{par.natural}{contains three parameters in AR(1) model. The first one is the stationary mean, the second is the AR coefficient, and the third is stationary variance.}

\item{yy}{the data.}

\item{mm}{the Monte Carlo sample size.}

\item{setseed}{the seed number.}

\item{resample}{the logical value indicating for resampling.}
}
\value{
The function returns the log-likelihood of the data.
}
\description{
The function implements the sequential Monte Carlo method using sequential importance sampling for stochastic volatility models.
}
\references{
Tsay, R. and Chen, R. (2018). Nonlinear Time Series Analysis. John Wiley & Sons, New Jersey.
}
