% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flow.R
\name{write_level_flow_simulation}
\alias{write_level_flow_simulation}
\title{Add flow through each network level to the simulation animation.}
\usage{
write_level_flow_simulation(
  network,
  level_throughput,
  res,
  x_coord,
  text_prefix = "Flow: ",
  adjustment = 1
)
}
\arguments{
\item{network}{data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`}

\item{level_throughput}{data.frame containing the maximum flow in network levels with the columns `level_id`, `throughtput`}

\item{res}{list containing the simulated data needed to animate the simulation}

\item{x_coord}{numeric, this is the `x` coordinate.}

\item{text_prefix}{string, prefix that should be added in front of values. Default is "Flow: ".}

\item{adjustment}{numeric, throughput is multiplied by this value. It can be used in case that the definition of network is in seconds and desired througput should be in hours.}
}
\value{
list containing the simulated data needed to animate the simulation
}
\description{
This function writes the maximum flow through each level
to the left side of the animation graph.
}
\examples{
network <- create_source(set = c(3,2,5), weight = 2.5) \%>\%
add_level_set(weight = c(2,4,1), previous_set = c(3,2,1),set = c(2,3, 1)) \%>\%
  add_level_set(weight = c(2,2,2),previous_set = c(1,2,3),set = c(1,1,1)) \%>\%
  add_fully_connected(weight = 5,set = 4) \%>\%
  add_exit()

level_throughput <- get_level_throughput(network, constrain = FALSE)
res <- simulate_flow(network)
res <- write_level_flow_simulation(network,
                        level_throughput,
                        res,
                        -6.5)
animate_simulation(network,
                   res[[1]],
                   res[[2]],
                  snap = 0)

}
