% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flow.R
\name{write_node_flow}
\alias{write_node_flow}
\title{Add flow info of each node to the simulation animation.}
\usage{
write_node_flow(
  network,
  node_load,
  res,
  input_people = TRUE,
  output_people = TRUE,
  node_flow = TRUE
)
}
\arguments{
\item{network}{data.frame with the definition of edges by the columns `from`, `to`, `weight`, `level`, `set`}

\item{node_load}{data.frame with the definition of nodes by the columns `id`, `input_people`, `output_people`, `node_load`}

\item{res}{list containing the simulated data needed to animate the simulation}

\item{input_people}{bool indicating whether to write maximum possible input to the node  into the graph}

\item{output_people}{bool indicating whether to write maximum possible output from the node into the graph}

\item{node_flow}{bool indicating whether to write node load at maximum input and output into the graph}
}
\value{
list containing the simulated data needed to animate the simulation
}
\description{
This function writes the maximum input, output, and node load of each node
into the graph during animation.
}
\examples{
network <- create_source(set = c(3,2,5), weight = 2.5) \%>\%
  add_level_set(weight = c(2,4,1), previous_set = c(3,2,1),set = c(2,3, 1)) \%>\%
  add_level_set(weight = c(2,2,2),previous_set = c(1,2,3),set = c(1,1,1)) \%>\%
  add_fully_connected(weight = 5,set = 4) \%>\%
  add_exit()

node_load <- get_node_load(network)
res <- simulate_flow(network)
res <- write_node_flow(network, node_load, res)
animate_simulation(network,
                   res[[1]],
                   res[[2]],
                  snap = 0)
}
