% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rankQ}
\alias{rankQ}
\title{Rank-Based Portmanteau Tests}
\usage{
rankQ(zt, lag = 10, output = TRUE)
}
\arguments{
\item{zt}{time series.}

\item{lag}{the maximum lag to calculate the test statistic.}

\item{output}{a logical value for output. Default is TRUE.}
}
\value{
\code{rankQ} function outputs the test statistics and p-values for Portmanteau tests, and returns a list with components:
\item{Qstat}{test statistics.}
\item{pv}{p-values.}
}
\description{
Performs rank-based portmanteau statistics.
}
\examples{
y=rnorm(1000)
rankQ(y,10,output=TRUE)
}
